refactor: update MongoDB connection path to db.config.js for consistency
